Rework inheritance for UWP
